Separate functions for extracting webpage data and computing readability
