+++ TextExtractionFunc/ExtractText/index.js
@@ -1,7 +1,4 @@
 var Mercury = require('@postlight/mercury-parser');
-var Readability = require('text-readability');
-var extractor = require("./extractMetrics.js");
-var natural = require("natural")
 
 /**
  * Parse api request
@@ -21,18 +18,10 @@ function parseRequestInput(request, context) {
     var url = request.body["url"];
     var isUrlValid = isValidUrl(url, context);
 
-    context.log("Checking for valid content");
-    var content = request.body["content"];
-    var isContentValid = false;
-    if (content != null && content.trim() != "")
-    {
-        isContentValid = true;
-    }
-
-    if (!isUrlValid && !isContentValid) {
+    if (!isUrlValid) {
         return {
             status: 400,
-            body: "Please enter valid url or content."
+            body: "Please enter valid url."
         }
     }
 
@@ -78,68 +67,17 @@ async function getParsedText(parsedInput, context) {
     // use mercury to get the web page text
     parsedTextResult = {};
     var url = parsedInput.url;
-    if (parsedInput.isUrlValid) {
-        try {
-            context.log("Calling mercury service to parse text from url.");
-            parsedTextResult = await Mercury.parse(url, {contentType: 'text'});
-            context.log("Successfully retrieved parsed web page.");
-        }
-        catch (error) {
-            context.log(error);
-            throw error;
-        }
+    try {
+        context.log("Calling mercury service to parse text from url.");
+        parsedTextResult = await Mercury.parse(url, {contentType: 'text'});
+        context.log("Successfully retrieved parsed web page.");
     }
-    else {
-        context.log("Url not specified. Parsing user entered text");
-        text = parsedInput.content;
-        if (text == null || text.trim() == "")
-        {
-            throw `Could not parse text content`
-        }
-        parsedTextResult["content"] = text;
+    catch (error) {
+        context.log(error);
+        throw error;
     }
 
     return parsedTextResult;
-}
-
-/**
- * split the text into sentences and then rejoin it to 
- * get rid of newlines and other separators
- * The documentation for this tokenizer can be found here
- * https://github.com/NaturalNode/natural#tokenizers
- * This is from the 'natural' npm package
- * @param {string} text the text to split
- * @param {*} context the azure function context
- */
-function splitSentences(text, context) {
-    context.log("parsing sentences...");
-
-    // Right now, the SentenceTokenizerNew is better, but not as robust as
-    // SentenceTokenizer, so we will try to use sentenceTokenizerNew and
-    // then default to SentenceTokenizer if there are errors
-    try {
-        var tokenizer = new natural.SentenceTokenizerNew();
-        sentences = tokenizer.tokenize(text);
-    }
-    catch (error) {
-        context.log(`Unable to parse with SentenceTokenizerNew because of ${error}`);
-        tokenizer = new natural.SentenceTokenizer();
-        sentences = tokenizer.tokenize(text);
-    }
-
-    // If the split sentence does not end in punctuation, then add a period because
-    // we need to be able to distinguish sentences after smushing them together.
-    for (var i=0; i<sentences.length; i++) {
-        if (sentences[i].match("([^.!?]*[.!?])+") == null) {
-            context.log(`Sentence without ".": ${sentences[i]}`);
-            sentences[i] = sentences[i] + ".";
-        }
-    }
-    // When the sentences are split, they still contain the punctuation
-    // so we only need to add a space when joining.
-    var joinedSentences = sentences.join(" ");
-
-    return joinedSentences;
 }
 
 // The main function that is executed in the azure function
@@ -165,23 +103,6 @@ module.exports = async function (context, request) {
     var content = splitSentences(parsedTextResult.content, context);
     // var content = parsedTextResult.content;
 
-    // Compute a bunch of readability metrics
-    var syllableCount = Readability.syllableCount(content, lang='en-US');
-    var lexiconCount = Readability.lexiconCount(content, removePunctuation=true);
-    var sentenceCount = Readability.sentenceCount(content);
-    var difficultWords = Readability.difficultWords(content);
-    var averageSentenceLength = Readability.averageSentenceLength(content);
-    var lixReadabilityIndex = Readability.lix(content);
-    var fleschEase = Readability.fleschReadingEase(content);
-    var fleschKincaidGrade = Readability.fleschKincaidGrade(content);
-    var colemanLiauIndex = Readability.colemanLiauIndex(content);
-    var automatedReadabilityIndex = Readability.automatedReadabilityIndex(content);
-    var daleChallReadabilityScore = Readability.daleChallReadabilityScore(content);
-    var linsearWriteIndex = Readability.linsearWriteFormula(content);
-    var gunningFogIndex = Readability.gunningFog(content);
-    var smogIndex = Readability.smogIndex(content);
-    var overallScore = Readability.textStandard(content);
-
     // Contruct the new response object
     var responseBody = {
         author: parsedTextResult.author || "",
@@ -198,21 +119,6 @@ module.exports = async function (context, request) {
         total_pages: parsedTextResult.total_pages || 0,
         url: parsedTextResult.url || "",
         word_count: parsedTextResult.word_count || 0,
-        syllable_count: syllableCount || 0,
-        lexicon_count: lexiconCount || 0,
-        sentence_count: sentenceCount || 0,
-        average_sentence_length: averageSentenceLength || 0,
-        lix_readability_index: lixReadabilityIndex || 0,
-        flesch_ease: fleschEase || 0,
-        fleschkincaid_grade: fleschKincaidGrade || 0,
-        coleman_liau_index: colemanLiauIndex || 0,
-        automated_readability_index: automatedReadabilityIndex || 0,
-        dale_chall_readability_score: daleChallReadabilityScore || 0,
-        difficult_words: difficultWords || 0,
-        linsear_write_index: linsearWriteIndex || 0,
-        gunning_fog_index: gunningFogIndex || 0,
-        smog_index: smogIndex || 0,
-        overall_score: overallScore || ""
     };
 
     context.log("function executed successfully");
