+++ TextExtractionFunc/host.json
@@ -4,5 +4,5 @@
         "id": "Microsoft.Azure.Functions.ExtensionBundle",
         "version": "[1.*, 2.0.0)"
     },
-    "functions": [ "ExtractText" ]
+    "functions": [ "ExtractText", "ComputeReadability" ]
 }
